Initial commit
Got a lot done but thought it was time to commit on the chance I lose it. There are several components. About one per page and the router is hooked up to allow it to work. Also implemented emailjs to have the ability to send emails if the person would like to contact me through the website.